requirements.txt
pyaml
discord.py
plotly
yattag
